refactor layout for copilot lsp
