+++ src/wasm-lib/kcl/src/server/copilot/mod.rs
@@ -1,0 +1,1 @@
+//! The copilot lsp server for ghost text.
